requirements.txt
Flask
pprint
logzero
pyjwt
requests

